{{cookiecutter.repo_name}}/requirements/local.txt
-r testing.txt

{% if cookiecutter.with_documentation == 'yes' %}Sphinx{% endif %}

django-debug-toolbar
django-debug-toolbar-template-timings
django-debug-toolbar-line-profiler
django-debug-toolbar-template-profiler
